move part of the code into automation crate
